[dataflow-core] Add (de)serializer singleton to allow dynamic type.
